refactor: file structure in translations
