+++ lib/config/translations_config.dart
