+++ lib/services/translations_service/business/flush_cache_drift_use_cases.dart
@@ -1,9 +1,8 @@
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
 import "../../../config/translations_config.dart";
-import "../local/database/translation_db_singleton.dart";
+import "../data/data_source/local/database/translation_db_singleton.dart";
 
-// TODO(simon-the-shark): @tomasz-trela move this
 extension FlushCacheDriftUseCases on WidgetRef {
   Future<void> flushOldData() async {
     final db = watch(translationsDatabaseSingletonProvider);
