+++ lib/services/translations_service/widgets/remove_old_translations.dart
@@ -3,8 +3,8 @@ import "dart:async";
 import "package:flutter/widgets.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 
-import "../../../api_base/cache/remote_flush/business/flush_cache_use_cases.dart";
 import "../../../hooks/use_effect_on_init.dart";
+import "../business/flush_cache_drift_use_cases.dart";
 
 class RemoveOldTranslations extends HookConsumerWidget {
   const RemoveOldTranslations({required this.child, super.key});
@@ -15,8 +15,8 @@ class RemoveOldTranslations extends HookConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     useEffectOnInit(() {
       unawaited(
-        Future.microtask(ref.flushAllCacheIfNeeded),
-      ); // TODO(simon-the-shark): @tomasz-trela flush translations cache instead of graphql cache
+        Future.microtask(ref.flushOldData),
+      );
       return null;
     });
     return child;
